add transition day

+++ Task7/src/databean/Position.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class Position  implements Serializable{
 	private int customer_id;
 	private int fund_id;
-	private int shares;
+	private long shares;
 	
 	public int getCustomer_id() {
 		return customer_id;
@@ -19,10 +19,10 @@ public class Position  implements Serializable{
 	public void setFund_id(int fund_id) {
 		this.fund_id = fund_id;
 	}
-	public int getShares() {
+	public long getShares() {
 		return shares;
 	}
-	public void setShares(int shares) {
+	public void setShares(long shares) {
 		this.shares = shares;
 	}
